Added TestGeneratorsHelper and test of parent option for ModelGenerator
