+++ test/unit/test_config_generator.rb
@@ -1,13 +1,5 @@
-require 'rails/generators'
-require 'rails/generators/mongo_mapper/config/config_generator'
-
 class ConfigGeneratorTest < Rails::Generators::TestCase
-  tests MongoMapper::Generators::ConfigGenerator
-
-  destination File.expand_path('../tmp', File.dirname(__FILE__))
-
-  setup :prepare_destination
-  teardown :cleanup_destination_root
+  include TestGeneratorsHelper
 
   test 'mongo.yml are properly created' do
     run_generator
@@ -18,11 +10,5 @@ class ConfigGeneratorTest < Rails::Generators::TestCase
     run_generator %w{dummy}
     assert_file 'config/mongo.yml', /dummy/
   end
-
-  protected
-
-    def cleanup_destination_root
-      rm_rf(destination_root)
-    end
 
 end
